requirements.txt
openai
elasticsearch
requests
pymupdf
jieba
python-dotenv
langchain
langchain-community
tiktoken
langchain-text-splitters
